Disable ALS because of problematic userspace
ALS should be enabled only in distros with proper auto-brightness support.

The issue mostly stems from not being able to control that option in login screen.
And not being able to do that, causes the login backlight value to override DT or DE value.

Other distros that don't use the sensor for backlight or user can properly control it in login AND DE are fine.

Examples of proper support:
- Android (fully controlled and apps support)
- Lakka (support in emulators for detecting environmental light conditions)
- Possibly some linux DEs