dc: Disable WinD on DispA
Some userspace applications can try and use Window D which is limited compared to A/B/C.
That can create all shorts of issues and glitches on display.

So disable access to it.

Additionally, remove access to it for DispB also.
It was never supported and it causes many issues with the display driver on init.